+++ .gitignore
@@ -1,0 +1,2 @@
+.log
+data analyst projs.code-workspace
